This is the initial upload of the AS3 AIR port of Flubble, by Austin of Two States Away (http://www.twostatesaway.com).
This upload contains the core display of the class stub creation system. It features full auto complete for AS3 class types, scrollable window, among other things.

The next step in the process is to add modules for saving class stubs in AS3, XML, JSON, MXML, and Flubber project files.

Over the course of development, I'll also be adding functionality for Pastebin, Dropbox, SVN, GIT and other things that pop into my head.

This is still a work in progress, and not ready for release. I'm still waiting on a response from the original creator giving the go ahead with the project.

Signed-off-by: Jerome Doby <xaerodegreaz@gmail.com>